+++ src/components/NoteCard.tsx
@@ -3,27 +3,46 @@
 import toast from "react-hot-toast";
 import { Preview } from "./preview";
 import { BsTrash } from "react-icons/bs";
+import Link from "next/link";
 
 interface NoteCardProps {
   note_id: string;
   title: string;
   note: string;
+  topicId: string;
 }
 
-const NoteCard = ({ note_id, title, note }: NoteCardProps) => {
+const NoteCard = ({ note_id, title, note, topicId }: NoteCardProps) => {
   const handleDelete = async () => {
     const noteId = note_id;
-    const response = await fetch(`/api/notes/noteId?noteId=${noteId}`, {
-      method: "DELETE",
-    });
-    const data = await response.json();
-    if (data) {
-      toast.success("Note Deleted");
-      window.location.reload();
-    } else {
-      toast.error("Error deleting note");
+    const insertDeletedNote = await fetch(
+      `/api/notes/noteId?noteId=${noteId}`,
+      {
+        method: "POST",
+        body: JSON.stringify({
+          id: note_id,
+          title: title,
+          content: note,
+          topicId: topicId,
+        }),
+      }
+    );
+    // DO SOMETHING If DATA RETURNS.. CONTINUE WITH DELETE
+    const insertedNote = await insertDeletedNote.json();
+    if (insertedNote) {
+      const response = await fetch(`/api/notes/noteId?noteId=${noteId}`, {
+        method: "DELETE",
+      });
+      const data = await response.json();
+      if (data) {
+        toast.success("Note Deleted");
+        window.location.reload();
+      } else {
+        toast.error("Error deleting note");
+      }
     }
   };
+
   return (
     <div
       className="rounded-2xl bg-fuchsia-900 text-white h-[180px] shadow-xl p-4 relative cursor-pointer z-20"
@@ -35,11 +54,12 @@ const NoteCard = ({ note_id, title, note }: NoteCardProps) => {
           onClick={handleDelete} // a delete by note id. remember to add to recently deleted for all delete functions
         />
       </div>
-
-      <h3>{title}</h3>
-      <div className="pb-2">
-        <Preview value={note} />
-      </div>
+      <Link href={`/${topicId}/notes/${note_id}`}>
+        <h3>{title}</h3>
+        <div className="pb-2">
+          <Preview value={note} />
+        </div>
+      </Link>
     </div>
   );
 };
